config: Add check_set for checking if an operation would be Ok
`check_set` is an associated function on the type that checks if a given
path exists, and whether the value at that path function can be
deserialized from the given string.

Use-case: Important for later use in gsh, because we need to check if we
can set something before issuing a command that actually sets something.
Gsh does not run synchronously so it can not run set and check the
return value.